9140873414385988574
ManifestID: 9140873414385988574
Date: 25 November 2020 – 21:12:44 UTC
Version: CS-442154

+++ Assembly-CSharp/WorldGenSpawner.cs
@@ -123,7 +123,7 @@ public class WorldGenSpawner : KMonoBehaviour
 			Grid.Spawnable[i] = 0;
 		}
 		float num = 16.5f;
-		float num2 = 18f;
+		int num2 = 18;
 		Vector2I baseStartPos = SaveGame.Instance.worldGen.SpawnData.baseStartPos;
 		GridVisibility.Reveal(baseStartPos.x, baseStartPos.y, num2, num);
 	}
